+++ CHANGELOG.md
@@ -14,7 +14,10 @@ please take a look at related PRs and issues and see if the change affects you.
 
 ## [Unreleased]
 
+- UnorderedGroup matching made deterministic ([#73])
 
+
+[#73]: https://github.com/textX/Arpeggio/issues/73
 [Unreleased]: https://github.com/textX/Arpeggio/compare/v1.9.2...HEAD